requirements.txt
Django==5.1.6
django_multiselectfield==1.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
firebase_admin==7.1.0
Requests==2.32.5
